+++ src/operators/tensor/implementations/tensor_fp64x64.cairo
@@ -438,7 +438,7 @@ impl FP64x64Tensor of TensorTrait<FP64x64> {
     }
 
     fn gather_elements(
-        self: @Tensor<FP64x64>, indices: Tensor<usize>, axis: Option<usize>
+        self: @Tensor<FP64x64>, indices: Tensor<i32>, axis: Option<i32>
     ) -> Tensor<FP64x64> {
         math::gather_elements::gather_elements(self, indices, axis)
     }
